// src/models/IoK8sApiNetworkingV1NetworkPolicy/types.ts
import { IoK8sApiNetworkingV1NetworkPolicySpecModelOptions as IoK8sApiNetworkingV1NetworkPolicySpec } from '../IoK8sApiNetworkingV1NetworkPolicySpec';
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiNetworkingV1NetworkPolicyModelOptions {
  apiVersion?: string;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  spec?: IoK8sApiNetworkingV1NetworkPolicySpec;
}